+++ src/pages/_app.js
@@ -14,8 +14,27 @@ import '../assets/css/spinner.scss'
 import 'tippy.js/dist/tippy.css';
 import 'tippy.js/themes/material.css';
 
+function getCookie(cookieString, cname) {
+    var name = cname + "=";
+    var decodedCookie = decodeURIComponent(cookieString);
+    var ca = decodedCookie.split(';');
+    for(var i = 0; i <ca.length; i++) {
+      var c = ca[i];
+      while (c.charAt(0) == ' ') {
+        c = c.substring(1);
+      }
+      if (c.indexOf(name) == 0) {
+        return c.substring(name.length, c.length);
+      }
+    }
+    return false;
+}
+
 function Disq({ Component, pageProps }) {
     React.useEffect(() => {
+        const cookie = getCookie(document.cookie, "disq_theme")
+        if(cookie) document.querySelector("html").classList = `theme-${cookie}`
+
         if(document.getElementById("curtain")){
             setTimeout(() => {
                 document.getElementById("curtain").style.display = "none";