+++ src/pages/dashboard/index.js
@@ -155,41 +155,26 @@ function QuickActions() {
 function Changelog() {
     return (
         <div className="land_cl">
-            <h2 className="nomargin_left land_header">Changelog</h2>
-
-            <h4 className="cl_version nomargin_left land_header">v1.4.2</h4>
+            <h2 className="nomargin_left land_header">
+                <span className="cl_version">v1.5.0</span> -
+                What's new?
+            </h2>
             <div className="changelog">
                 <p>
-                + Collapsable sidebar<br/>
-                + Starting to optimise a few pages for mobile<br/>
-                + Redesigned Files and Short URL pages<br/>
-                + GitHub Actions CI<br/>
+                + We've merged with the Histacom team! Join our <a href="https://discord.gg/3eRxFpdK8z">Discord</a> for more info.<br/>
+                + Light and Ultradark themes<br/>
+                + Disq Plus (support us and get perks!)<br/>
+                + Vanity Short URLs for Plus users<br/>
+                + 70mb Upload limit for Plus users<br/>
+                + Linkpages (showcase your awesome social media)<br/>
+                + Home button on left to access this page easier<br/>
+                + Loading spinner to make the site look more polished<br/>
+                + Domain selecter in Upload page<br/>
+                + File upload limit displayed in Upload page<br/>
+                + Removed Changelog on Homepage and replaced with "What's new" (how meta)<br/>
                 <br/>
-                - Removed Censor links toggle<br/>
-                </p>
-            </div>
-
-            <h4 className="cl_version nomargin_left land_header">v1.4.1</h4>
-            <div className="changelog">
-                <p>
-                = Disq should now run faster as it turns out i was not building for production. my bad!
-                </p>
-            </div>
-
-            <h4 className="cl_version nomargin_left land_header">v1.4.0</h4>
-            <div className="changelog">
-                <p>
-                + Redesigned landing page with quick upload<br/>
-                + New greeting messages<br/>
-                + Tooltips everywhere!<br/>
-                + Server Side Rendering for Short URLs (NextJS)<br/>
-                + Social media embeds when linking Disq<br/>
-                + Redesigned Files and Short URL pages<br/>
+                = Fixed NaN bug in Short URLs page<br/>
                 <br/>
-                = Made loading pages smoother in the Files manager and Short URL manager.<br/>
-                <br/>
-                - Cringed at .php, .cpp, .c and .h files uploaded<br/>
-                - Removed easter egg with Request Verification in the settings page. It'll be back in a newer update!<br/>
                 </p>
             </div>
         </div>